Implement __LOCATION__ macro
Implement __LOCATION__ macro to give macros access to the line number in
the context they were called.  This is expanded during the variable
resolution step after each macro is executed.

* Ensure that line number and filename information is preserved in the AST
when running from the REPL, as the REPL expands and evals the AST in a
separate task from where it's parsed.  This allows current_location() to
work properly at the REPL.