Add current_lineno to TLS for macro location tracking
Here we pass the line number (tracked inside macroexpand.scm) into
fl_invoke_julia_macro and stash it on the thread local storage, from
which it can be retrieved inside a macro using the current_location()
function.

This is a bit ugly but gets the job done - it seems like the only
alternative would be to change the macro calling convention to insert an
implicit context parameter (an ABI change only, but much more
disruptive).